+++ config/default.go
@@ -6,7 +6,6 @@ import (
 	"github.com/metafates/mangal/color"
 	"github.com/metafates/mangal/constant"
 	"github.com/metafates/mangal/style"
-	"github.com/samber/lo"
 	"github.com/spf13/viper"
 )
 
@@ -38,22 +37,22 @@ func (f *Field) typeName() string {
 	}
 }
 
-// Json returns the config as json string
-func (f *Field) Json() string {
+func (f *Field) MarshalJSON() ([]byte, error) {
 	field := struct {
 		Key         string `json:"key"`
 		Value       any    `json:"value"`
+		Default     any    `json:"default"`
 		Description string `json:"description"`
 		Type        string `json:"type"`
 	}{
 		Key:         f.Key,
-		Value:       f.Value,
+		Value:       viper.Get(f.Key),
+		Default:     f.Value,
 		Description: f.Description,
 		Type:        f.typeName(),
 	}
 
-	output := lo.Must(json.Marshal(field))
-	return string(output)
+	return json.Marshal(field)
 }
 
 // Pretty format field as string for further cli output