+++ constant/duration.go
@@ -1,5 +1,0 @@
-package constant
-
-import "time"
-
-const Forever = time.Duration(1<<63 - 1)